requirements.txt
flask
numpy
pillow
gdown
tensorflow==2.17.1  # or 2.19.0 — works inside Docker
gunicorn
